+++ xtask/src/device_path/mod.rs
@@ -30,8 +30,9 @@ fn gen_code_as_string(groups: &[NodeGroup]) -> Result<String> {
         use crate::proto::network::IpAddress;
         use crate::table::boot::MemoryType;
         use core::mem::{size_of, size_of_val};
-        use core::ptr::{self, addr_of};
+        use core::ptr::addr_of;
         use core::{fmt, slice};
+        use ptr_meta::{Pointee, PtrExt};
 
         #(#packed_modules)*